+++ src/components/formSections/AmendmentFormInfo.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { Field } from 'redux-form';
+import { Field } from 'react-final-form';
+import { requiredValidator } from '@folio/stripes-erm-components';
 
 import {
   Checkbox,
@@ -13,7 +14,7 @@ import {
   TextField,
 } from '@folio/stripes/components';
 
-import { required } from '../../util/validators';
+import WarnEndDate from '../WarnEndDate';
 
 class AmendmentFormInfo extends React.Component {
   static propTypes = {
@@ -21,25 +22,8 @@ class AmendmentFormInfo extends React.Component {
       statusValues: PropTypes.array,
       typeValues: PropTypes.array,
     }),
+    form: PropTypes.object,
   };
-
-  state = {
-    openEnded: false,
-  }
-
-  warnEndDate = (_value, allValues) => {
-    this.setState({ openEnded: allValues.openEnded });
-
-    if (allValues.openEnded && allValues.endDate) {
-      return (
-        <div data-test-warn-clear-end-date>
-          <FormattedMessage id="ui-licenses.warn.clearEndDate" />
-        </div>
-      );
-    }
-
-    return undefined;
-  }
 
   validateEndDate = (value, allValues) => {
     if (value && allValues.startDate && (allValues.openEnded !== true)) {
@@ -59,7 +43,7 @@ class AmendmentFormInfo extends React.Component {
   }
 
   render() {
-    const { data } = this.props;
+    const { data, form } = this.props;
 
     return (
       <div
@@ -77,7 +61,7 @@ class AmendmentFormInfo extends React.Component {
                   component={TextField}
                   placeholder={placeholder}
                   required
-                  validate={required}
+                  validate={requiredValidator}
                 />
               )}
             </FormattedMessage>
@@ -107,26 +91,35 @@ class AmendmentFormInfo extends React.Component {
             />
           </Col>
           <Col xs={10} md={5}>
-            <Field
-              backendDateStandard="YYYY-MM-DD"
-              id="edit-amendment-end-date"
-              name="endDate"
-              label={<FormattedMessage id="ui-licenses.prop.endDate" />}
-              component={Datepicker}
-              dateFormat="YYYY-MM-DD"
-              disabled={this.state.openEnded}
-              validate={this.validateEndDate}
-              warn={this.warnEndDate}
-            />
+            <Field name="endDate" validate={this.validateEndDate}>
+              {(props) => {
+                const formState = form.getState();
+                const { values = {} } = formState;
+                return (<Datepicker
+                  backendDateStandard="YYYY-MM-DD"
+                  id="edit-amendment-end-date"
+                  label={<FormattedMessage id="ui-licenses.prop.endDate" />}
+                  dateFormat="YYYY-MM-DD"
+                  disabled={values.openEnded}
+                  onBlur={props.input.onBlur}
+                  input={props.input}
+                  meta={props.meta}
+                />);
+              }}
+            </Field>
           </Col>
           <Col xs={2} style={{ paddingTop: 20 }}>
-            <Field
-              id="edit-amendment-open-ended"
-              name="openEnded"
-              label={<FormattedMessage id="ui-licenses.prop.openEnded" />}
-              component={Checkbox}
-              type="checkbox"
-            />
+            <Field name="openEnded" type="checkbox">
+              {props => {
+                return (<Checkbox
+                  id="edit-amendment-open-ended"
+                  checked={props.input.value}
+                  label={<FormattedMessage id="ui-licenses.prop.openEnded" />}
+                  onChange={props.input.onChange}
+                  type="checkbox"
+                />);
+              }}
+            </Field>
           </Col>
         </Row>
         <Row>
@@ -144,6 +137,7 @@ class AmendmentFormInfo extends React.Component {
             </FormattedMessage>
           </Col>
         </Row>
+        <WarnEndDate mutators={form.mutators} />
       </div>
     );
   }